[#9] refactor: 확대이미지 스타일 분리
